+++ src/main.rs
@@ -1,19 +1,13 @@
 use rayon::iter::{IntoParallelIterator, ParallelIterator};
-use std::{
-    env,
-    ffi::OsStr,
-    path::{Path, PathBuf},
-};
-use walkdir::WalkDir;
+use std::
+    env
+;
 
-mod image_actions;
+use crate::image_actions::{find_margins::find_margins, crop::crop};
+
 mod misc;
-
-use image_actions::find_empty_space::find_margins;
-
-pub fn load_image(path: &Path) -> image::DynamicImage {
-    image::open(path).expect("Failed to open image")
-}
+mod image_actions;
+mod filesystem;
 
 fn main() {
     let args: Vec<String> = env::args().collect();
@@ -26,7 +20,7 @@ fn main() {
     let input_directory = args[1].as_str();
     let output_directory = args[2].as_str();
 
-    let images_result = find_images(input_directory);
+    let images_result = filesystem::find_images(input_directory);
     let images = images_result.expect("Failed to list images");
     println!("Found {} images", images.len());
 
@@ -36,9 +30,9 @@ fn main() {
     println!("Start processing");
     images.into_par_iter().for_each(|file| {
         println!("Processing {}", file.display());
-        let mut img = load_image(&file);
+        let mut img = filesystem::load_image(&file);
         let margins = find_margins(&mut img, 100.0);
-        let cropped_img = image_actions::crop(
+        let cropped_img = crop(
             &mut img,
             margins.0 .0,
             margins.0 .1,
@@ -55,21 +49,3 @@ fn main() {
     });
 }
 
-fn find_images<P: AsRef<Path>>(path: P) -> Result<Vec<PathBuf>, std::io::Error> {
-    let mut images = Vec::new();
-
-    for entry in WalkDir::new(path.as_ref())
-        .into_iter()
-        .filter_map(|e| e.ok())
-    {
-        let path = entry.path();
-        if path.is_file() && path.extension() == Some(OsStr::new("jpg"))
-            || path.extension() == Some(OsStr::new("jpeg"))
-            || path.extension() == Some(OsStr::new("png"))
-        {
-            images.push(path.to_path_buf());
-        }
-    }
-
-    Ok(images)
-}
